requirements.txt
Matplotlib
Pillow
python-dotenv
PyYAML
scikit-learn
tensorflow
tensorflow-hub
valohai-cli
valohai-utils